+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## [Unreleased]
 
 - feat: allow studio infra to accept edge lambdas for CloudFront behaviors
+- feat: exchange cognito oauth codes at edge and set auth cookie